+++ system-tests/services/test-apps.js
@@ -1,3 +1,5 @@
+require('../_support/utils/ServicesUtil');
+
 describe('Services', function () {
 
   /**
