+++ system-tests/services/test-pods.js
@@ -1,3 +1,5 @@
+require('../_support/utils/ServicesUtil');
+
 describe('Services', function () {
 
   /**
